Removed include/klee/util and moved header files to appropriate places
