+++ include/klee/Expr/ArrayExprOptimizer.h
@@ -17,9 +17,9 @@
 #include <utility>
 #include <vector>
 
+#include "klee/ADT/Ref.h"
 #include "klee/Expr/Expr.h"
 #include "klee/Expr/ExprHashMap.h"
-#include "klee/util/Ref.h"
 
 namespace klee {
