+++ include/klee/Expr/ArrayExprRewriter.h
@@ -14,8 +14,8 @@
 #include <map>
 #include <vector>
 
+#include "klee/ADT/Ref.h"
 #include "klee/Expr/Expr.h"
-#include "klee/util/Ref.h"
 
 namespace klee {
