+++ include/klee/Expr/ExprSMTLIBPrinter.h
@@ -14,7 +14,7 @@
 #include "klee/Expr/Constraints.h"
 #include "klee/Expr/Expr.h"
 #include "klee/Solver/Solver.h"
-#include "klee/util/PrintContext.h"
+#include "klee/Support/PrintContext.h"
 
 #include <map>
 #include <set>
