+++ lib/Expr/ArrayExprOptimizer.cpp
@@ -9,6 +9,7 @@
 
 #include "klee/Expr/ArrayExprOptimizer.h"
 
+#include "klee/ADT/BitArray.h"
 #include "klee/Config/Version.h"
 #include "klee/Expr/ArrayExprRewriter.h"
 #include "klee/Expr/ArrayExprVisitor.h"
@@ -17,7 +18,6 @@
 #include "klee/Expr/ExprBuilder.h"
 #include "klee/OptionCategories.h"
 #include "klee/Support/ErrorHandling.h"
-#include "klee/util/BitArray.h"
 
 #include <llvm/ADT/APInt.h>
 #include <llvm/Support/Casting.h>
